requirements.txt
langchain
langchain-community
langchain-openai
datetime
openai
python-docx
streamlit
sendgrid
pydub
streamlit-audiorecorder
st-annotated-text